requirements.txt
Django==3.1.14
django-debug-toolbar==3.2.2
djangorestframework==3.12.2
pytz==2021.3
django-filter==2.4.0
django-notifications-hq==1.6.0
boto3==1.20.26
django-storages==1.12.3
python-memcached==1.59
